projects/GAA-GCP-ILS-4/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-ils"
version = "1.0.0"
description = "Graph Coloring Problem with Iterated Local Search"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "GCP-ILS Project Team"},
]
keywords = ["optimization", "graph-coloring", "combinatorial-optimization", "metaheuristics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
]

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-watch>=4.2.0",
    "mypy>=0.950",
    "pylint>=2.9.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]
docs = [
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/gcp-ils"
Documentation = "https://gcp-ils.readthedocs.io"
Repository = "https://github.com/your-org/gcp-ils.git"
"Bug Tracker" = "https://github.com/your-org/gcp-ils/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "operators*", "metaheuristic*", "utils*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.pylint.messages_control]
disable = [
    "C0330",  # Wrong hanging indentation before block
    "C0326",  # Bad whitespace
]

[tool.pylint.format]
max-line-length = 100
